Initial Receiver to provide caller with packets
The caller can implement Receiver and the implementation will be passed
packets after the application of FEC processing.

Doesn't currently report to the Receiver how many packets where lost /
recovered, but the intent is that this should be possible.

Also the implementation could do with some tidying up and hopefully
simplification.