Remove PacketRef/PacketRefMut abstractons
No point in having them after I stopped PacketRef implementing Clone.

This does mean that it's not possible for the FEC matrix to have a
readable copy of the packet and for the decoder to also pass a readable
copy to the application, which was my original intent.  For now the app
will only get packets once they exit the matrix.

May need to revisit this decision later, but for now having simpler code
is probably the best course to support evolution of the codebase.