We need recovered seq num, not received seq num
Also, add some assertions to try and catch inconsistencies between the
value in the packet headers in the buffer, and the Seq value that we are
for some reason carrying around through the API along side the buffer.

Should probably remove the extra Seq values except where actually
needed, at some point!